ClipforgeServer/requirements.txt
boto3>=1.39,<2.0
fastapi>=0.115,<1.0
httpx>=0.27,<1.0
pillow>=11.0,<12.0
pytest>=8.0,<9.0
python-multipart>=0.0.9,<1.0
uvicorn[standard]>=0.30,<1.0
